test: relocate integration tests
