+++ crates/pop-cli/Cargo.toml
@@ -44,12 +44,7 @@ contract = [
     "dep:sp-core",
     "dep:sp-weights",
 ]
-e2e_contract = []
-unit_contract = []
 parachain = [
     "dep:pop-parachains",
     "dep:dirs",
 ]
-e2e_parachain = []
-unit_parachain = []
-
