+++ crates/pop-contracts/src/build.rs
@@ -15,44 +15,3 @@ pub fn build_smart_contract(path: &Option<PathBuf>) -> anyhow::Result<String> {
 
 	Ok(formatted_result)
 }
-
-#[cfg(feature = "unit_contract")]
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use anyhow::{Error, Result};
-	use std::fs;
-
-	fn setup_test_environment() -> Result<tempfile::TempDir, Error> {
-		let temp_dir = tempfile::tempdir().expect("Could not create temp dir");
-		let temp_contract_dir = temp_dir.path().join("test_contract");
-		fs::create_dir(&temp_contract_dir)?;
-		crate::create_smart_contract("test_contract", temp_contract_dir.as_path())?;
-		Ok(temp_dir)
-	}
-
-	#[test]
-	fn test_contract_build() -> Result<(), Error> {
-		let temp_contract_dir = setup_test_environment()?;
-
-		build_smart_contract(&Some(temp_contract_dir.path().join("test_contract")))?;
-
-		// Verify that the folder target has been created
-		assert!(temp_contract_dir.path().join("test_contract/target").exists());
-		// Verify that all the artifacts has been generated
-		assert!(temp_contract_dir
-			.path()
-			.join("test_contract/target/ink/test_contract.contract")
-			.exists());
-		assert!(temp_contract_dir
-			.path()
-			.join("test_contract/target/ink/test_contract.wasm")
-			.exists());
-		assert!(temp_contract_dir
-			.path()
-			.join("test_contract/target/ink/test_contract.json")
-			.exists());
-
-		Ok(())
-	}
-}
