+++ crates/pop-contracts/src/up.rs
@@ -96,74 +96,3 @@ pub async fn instantiate_smart_contract(
 	let instantiate_result = instantiate_exec.instantiate(Some(gas_limit)).await?;
 	Ok(instantiate_result.contract_address.to_string())
 }
-
-#[cfg(feature = "unit_contract")]
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use crate::{build_smart_contract, create_smart_contract};
-	use anyhow::{Error, Result};
-	use std::fs;
-	use tempfile::TempDir;
-	use url::Url;
-
-	const CONTRACTS_NETWORK_URL: &str = "wss://rococo-contracts-rpc.polkadot.io";
-
-	fn generate_smart_contract_test_environment() -> Result<tempfile::TempDir, Error> {
-		let temp_dir = tempfile::tempdir().expect("Could not create temp dir");
-		let temp_contract_dir = temp_dir.path().join("test_contract");
-		fs::create_dir(&temp_contract_dir)?;
-		create_smart_contract("test_contract", temp_contract_dir.as_path())?;
-		Ok(temp_dir)
-	}
-	fn build_smart_contract_test_environment(temp_dir: &TempDir) -> Result<(), Error> {
-		build_smart_contract(&Some(temp_dir.path().join("test_contract")))?;
-		Ok(())
-	}
-
-	#[tokio::test]
-	async fn test_set_up_deployment() -> Result<(), Error> {
-		let temp_dir = generate_smart_contract_test_environment()?;
-		build_smart_contract_test_environment(&temp_dir)?;
-
-		let call_opts = UpOpts {
-			path: Some(temp_dir.path().join("test_contract")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			salt: None,
-		};
-		let result = set_up_deployment(call_opts).await?;
-		assert_eq!(result.url(), "wss://rococo-contracts-rpc.polkadot.io:443/");
-		Ok(())
-	}
-
-	#[tokio::test]
-	async fn test_dry_run_gas_estimate_instantiate() -> Result<(), Error> {
-		let temp_dir = generate_smart_contract_test_environment()?;
-		build_smart_contract_test_environment(&temp_dir)?;
-
-		let call_opts = UpOpts {
-			path: Some(temp_dir.path().join("test_contract")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "0".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			salt: None,
-		};
-		let instantiate_exec = set_up_deployment(call_opts).await;
-
-		let weight = dry_run_gas_estimate_instantiate(&instantiate_exec.unwrap()).await?;
-		assert!(weight.clone().ref_time() > 0);
-		assert!(weight.proof_size() > 0);
-
-		Ok(())
-	}
-}
